+++ parameter/ParameterBlockType.h
@@ -47,6 +47,6 @@ private:
     // Population
     virtual void populate(CElement* pElement) const;
     // Creating sub blocks with indexes
-    static std::string computeChildName(uint32_t uiChild);
+    static std::string computeChildName(size_t uiChild);
 };
